backend/requirements.txt
chess==1.9.4
pyserial==3.5
python-dotenv==1.0.0
